+++ src/main/kotlin/com/sbl/sulmun2yong/survey/controller/SurveyInfoController.kt
@@ -50,4 +50,10 @@ class SurveyInfoController(
         @RequestParam status: SurveyStatus?,
         @RequestParam(defaultValue = "LAST_MODIFIED") sortType: MySurveySortType,
     ): ResponseEntity<MyPageSurveysResponse> = ResponseEntity.ok(surveyInfoService.getMyPageSurveys(userId, status, sortType))
+
+    @GetMapping("/make-info/{surveyId}")
+    override fun getSurveyMakeInfo(
+        @PathVariable("surveyId") surveyId: UUID,
+        @LoginUser id: UUID,
+    ) = ResponseEntity.ok(surveyInfoService.getSurveyMakeInfo(surveyId = surveyId, makerId = id))
 }
