+++ src/main/kotlin/com/sbl/sulmun2yong/survey/controller/SurveyWorkbenchController.kt
@@ -6,7 +6,6 @@ import com.sbl.sulmun2yong.survey.dto.request.SurveySaveRequest
 import com.sbl.sulmun2yong.survey.service.SurveyWorkbenchService
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.DeleteMapping
-import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PatchMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
@@ -38,12 +37,6 @@ class SurveyWorkbenchController(
             surveyId = surveyId,
         ),
     )
-
-    @GetMapping("/{surveyId}")
-    override fun getSurveyMakeInfo(
-        @PathVariable("surveyId") surveyId: UUID,
-        @LoginUser id: UUID,
-    ) = ResponseEntity.ok(surveyWorkbenchService.getSurveyMakeInfo(surveyId = surveyId, makerId = id))
 
     @PatchMapping("/start/{surveyId}")
     override fun startSurvey(
