+++ src/main/kotlin/com/sbl/sulmun2yong/survey/controller/doc/SurveyWorkbenchApiDoc.kt
@@ -3,12 +3,10 @@ package com.sbl.sulmun2yong.survey.controller.doc
 import com.sbl.sulmun2yong.global.annotation.LoginUser
 import com.sbl.sulmun2yong.survey.dto.request.SurveySaveRequest
 import com.sbl.sulmun2yong.survey.dto.response.SurveyCreateResponse
-import com.sbl.sulmun2yong.survey.dto.response.SurveyMakeInfoResponse
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.DeleteMapping
-import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PatchMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
@@ -31,13 +29,6 @@ interface SurveyWorkbenchApiDoc {
         @LoginUser id: UUID,
         @RequestBody surveySaveRequest: SurveySaveRequest,
     ): ResponseEntity<Unit>
-
-    @Operation(summary = "설문 제작 정보 API")
-    @GetMapping("/{surveyId}")
-    fun getSurveyMakeInfo(
-        @PathVariable("surveyId") surveyId: UUID,
-        @LoginUser id: UUID,
-    ): ResponseEntity<SurveyMakeInfoResponse>
 
     @Operation(summary = "설문 시작 API")
     @PatchMapping("/start/{surveyId}")
